more efficient global bools
STOBOOL is inefficient since it contains 2 branches, and 2 layers
of indirection (read the global var containing a ptr, read 1st byte at
that ptr) and a NULL test. Removing the NULL test will make dmake crash
if a makefile does "OOODMAKEMODE = \n" which will set the global char *
to NULL, since Def_macro() changes "" to NULL.

M_VAR_CHAR is totally unused by dmake, and the only bit we can reuse in
highest nibble of M_* flags. This way all the M_VAR flags remain in the
same nibble. A bit_var is not a "STOBOOL" style bool. A bit_var is
interpreted as ""/NULL being false, everything
else is true, "n" is true for example with a bit_var but false with
STOBOOL style bools.

Reuse M_VAR_CHAR to make a more efficiency bool type for dmake.