optimize _drop_mac
my C profiler says this it is the 6th most called function, called 100Ks
of times. By setting the value in the struct to NULL, before calling free,
this C function becames tailcall-able or more likely to be inlined. Ptr
hp doesn't need to be saved across the C function free() by moving the NULL
assignment.