remove overhead of unimplemented .SYMBOL feature
Do_stat is very hot code (usually hitting the dir cache). This member
argument is from before git blame goes. The feature is unimplemented but
is legally parsable and then is a runtime error. In all the callers except
for Stat_target(),and then only in the A_SYMBOL branch, members is NULL.

Move the fatal error to A_SYMBOL branch and rid of passing NULL as an
argument in all other branches. For diagnostic reasons, DO_STAT macro
can't have its member arg removed.

Change Do_touch so it looks the same.