use strpbrk instead of textbook inefficient roll your own version
Most libc strpbrk'es are written in assembly and/or use CS voodoo with
lookup arrays to eliminate the inner loop. Dmake's version does neither.
searchchars is usually DirBrkStr from Filedir() and Basename().
This one commit cut the execution time of CacheStat with a cache hit, in
half according to my profiler.