factor out multiple Basename() calls + remove arg for really_dostat
All execution paths execute Basename, this is for smaller machine code
size.

There is no reason for the caller of really_dostat to allocate a 0x20 to
0x40 byte long stat buffer on C stack, only for it to NOT be used by
default, since the default is calling CacheStat() which doesn't take a
buf pointer. Since C stack is usually in CPU cache, having a slot of
usually unused C stack is bad. Now only if the buffer is needed will it be
allocated on C stack.