use MS CRT's malloc null check, dont do our own
MS CRT has a feature for malloc to check for null, and dispatch a
handler/callback of our choosing. Instead of MSCRT checking for null,
checking for the handler, and if no handler, return null, then dmake
check for null, just do the NULL check once in CRT instead of doing it
in ~30 callsites in dmake, but only on Win32, where this feature exists.
Because _set_new_handler is compiled with C++ in the CRT src and does
NOT have extern "C", we have to use a small wrapper compiland to make it
available to plain C code. With VC's -GL the small wrapper is optimized
away by auto-inlining. HEAP_GENERATE_EXCEPTIONS is a similar feature, but
CRT's new_mode isn't implemented with HEAP_GENERATE_EXCEPTIONS atleast for
VC 2003, although it could be if MS wanted to.