fix bug, temp file not flushed before used by recipie process
In commit
"dont call fflush for each line written when generating a tmp file" I
stopped flushing the fd of a temp file after every line, but that
introduced a new bug, since Close_temp was not always closing a fd (IDK if
that is intentional or not) even though it looked like it did. Change
Close_temp to fflush the fd if it wasn't fclose-d. The bug presented itself
as a temp file being empty when GCC read it, but examining the temp showed
it was filled correctly because it was filled as dmake was exiting after
processing the GCC non-zero exit.